requirements.txt
# Core dependencies
python-dotenv==1.0.0
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.2

# Deep learning
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
opencv-python==4.8.0.76

# Web interface
flask-wtf==1.1.1
wtforms==3.0.1

# Utilities
tqdm==4.66.1
requests==2.31.0